aura/requirements.txt
google-adk
gunicorn
fastapi
uvicorn